BS/requirements.txt
asgiref==3.6.0
certifi==2022.12.7
charset-normalizer==3.1.0
Django==3.2
django-bootstrap-modal-forms==2.2.0
django-bootstrap3==22.2
django-cors-headers==3.14.0
django-forms-bootstrap==3.1.0
django-widget-tweaks==1.4.12
djangorestframework==3.14.0
idna==3.4
importlib-metadata==2.1.3
mysqlclient==2.1.1
Pillow==9.4.0
pytz==2022.7.1
requests==2.28.2
sqlparse==0.4.3
typing_extensions==4.5.0
urllib3==1.26.15
zipp==3.15.0
